Update Find-PSServiceAccounts
The return of $AllServiceAccountsSPNs within the "If ($DumpSPNs)" statement seems to break functionality, if user chooses GetTGS too. The return is done, but GetTGS is never processed in my tests. When moving the return down after the "If ($GetTGS)" statement, the TGS retrieval seems to work.

Also, minor detail, the two if-statements, "If ($DumpSPNs)" and  "If ($GetTGS)" are now alike (without "-eq $True"), both are [swithch] type variables.